requirements.txt
django
python-dotenv
psycopg2-binary
requests
django-filter